some refactoring, move some code into separate components
