+++ packages/lesswrong/components/community/modules/LocalGroups.tsx
@@ -1,0 +1,198 @@
+import { Components, registerComponent, } from '../../../lib/vulcan-lib';
+import React from 'react';
+import { createStyles } from '@material-ui/core/styles';
+import * as _ from 'underscore';
+import { useMulti } from '../../../lib/crud/withMulti';
+import { Link } from '../../../lib/reactRouterWrapper';
+import { cloudinaryCloudNameSetting } from '../../../lib/publicSettings';
+
+const styles = createStyles((theme: ThemeType): JssStyles => ({
+  localGroups: {
+    display: 'grid',
+    gridTemplateColumns: '1fr 1fr',
+    marginTop: 20,
+    [theme.breakpoints.down('sm')]: {
+      gridTemplateColumns: '1fr',
+      marginLeft: -4,
+      marginRight: -4,
+    }
+  },
+  localGroupsList: {
+    height: 440,
+    overflowY: 'scroll',
+    [theme.breakpoints.down('sm')]: {
+      height: 'auto',
+    },
+  },
+  localGroup: {
+    height: 116,
+    borderBottomWidth: 1,
+    borderBottomStyle: 'solid',
+    borderColor: "rgba(0, 0, 0, 0.1)",
+    '&:last-of-type': {
+      borderBottom: 'none'
+    },
+    [theme.breakpoints.down('xs')]: {
+      height: 'auto'
+    },
+  },
+  mobileImg: {
+    display: 'none',
+    height: 160,
+    backgroundColor: '#e2f1f4',
+    justifyContent: 'center',
+    alignItems: 'center',
+    [theme.breakpoints.down('xs')]: {
+      display: 'flex'
+    },
+  },
+  localGroupContent: {
+    height: 115,
+    background: 'white',
+    backgroundRepeat: 'no-repeat',
+    backgroundPositionY: 'center',
+    padding: '16px 16px 16px 150px',
+    [theme.breakpoints.down('xs')]: {
+      height: 'auto',
+      backgroundImage: 'none !important',
+      paddingLeft: 16,
+      paddingBottom: 30
+    },
+  },
+  localGroupNameRow: {
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'baseline',
+  },
+  localGroupName: {
+    ...theme.typography.headline,
+    fontSize: 18,
+    display: '-webkit-box',
+    "-webkit-line-clamp": 2,
+    "-webkit-box-orient": 'vertical',
+    overflow: 'hidden',
+    marginBottom: 0
+  },
+  localGroupDistance: {
+    flex: 'none',
+    ...theme.typography.commentStyle,
+    color: theme.palette.primary.dark,
+    fontSize: 14,
+    marginLeft: 14
+  },
+  localGroupLocation: {
+    ...theme.typography.commentStyle,
+    color: "rgba(0, 0, 0, 0.7)",
+    fontSize: 14,
+    lineHeight: '1.5em',
+    display: '-webkit-box',
+    "-webkit-line-clamp": 2,
+    "-webkit-box-orient": 'vertical',
+    overflow: 'hidden',
+    marginTop: 8,
+  },
+  localGroupsMap: {
+    marginTop: 50,
+    [theme.breakpoints.down('sm')]: {
+      display: 'none'
+    },
+  },
+}))
+
+
+const LocalGroups = ({userLocation, classes}: {
+  userLocation: {
+    lat: number,
+    lng: number,
+    known: boolean,
+    loading: boolean,
+  },
+  classes: ClassesType,
+}) => {
+  const { CommunityMapWrapper, CloudinaryImage2 } = Components
+  
+  /**
+   * Calculates the distance between the query location and the given lat/lng, as the crow flies
+   *
+   * @param {number} lat - latitude
+   * @param {number} lng - longitude
+   * @returns {number}
+   */
+  const distance = (lat, lng) => {
+    if (!userLocation) return null
+    
+    const toRad = (num) => num * Math.PI / 180
+    
+    const dLat = toRad(lat - userLocation.lat)
+    const dLng = toRad(lng - userLocation.lng)
+    const a = (Math.sin(dLat/2) * Math.sin(dLat/2)) + (Math.sin(dLng/2) * Math.sin(dLng/2) * Math.cos(toRad(userLocation.lat)) * Math.cos(toRad(lat)))
+    return Math.round(2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a)) * 6371)
+  }
+
+  let groupsListTerms: LocalgroupsViewTerms = {}
+  groupsListTerms = userLocation.known ? {
+    view: 'nearby',
+    lat: userLocation.lat,
+    lng: userLocation.lng,
+  } : {
+    view: 'local',
+  }
+  
+  const { results } = useMulti({
+    terms: groupsListTerms,
+    collectionName: "Localgroups",
+    fragmentName: 'localGroupsHomeFragment',
+    fetchPolicy: 'cache-and-network',
+    nextFetchPolicy: "cache-first",
+    limit: 200,
+    skip: userLocation.loading
+  });
+  
+  const cloudinaryCloudName = cloudinaryCloudNameSetting.get()
+
+  return (
+    <div className={classes.localGroups}>
+      <div className={classes.localGroupsList}>
+        {results?.map(group => {
+          const rowStyle = group.bannerImageId ? {
+            backgroundImage: `linear-gradient(to right, transparent, white 140px), url(https://res.cloudinary.com/${cloudinaryCloudName}/image/upload/c_fill,g_custom,h_115,w_140,q_auto,f_auto/${group.bannerImageId})`
+          } : {
+            backgroundImage: 'url(https://res.cloudinary.com/cea/image/upload/c_pad,h_80,w_140,q_auto,f_auto/ea-logo-square-1200x1200__1_.png), linear-gradient(to right, #e2f1f4, white 140px)'
+          }
+          
+          return <div key={group._id} className={classes.localGroup}>
+            <Link to={`/groups/${group._id}`} className={classes.mobileImg}>
+              {group.bannerImageId ?
+                <CloudinaryImage2 height={160} width="100vw" objectFit="cover" publicId={group.bannerImageId} imgProps={{w:'600'}} /> :
+                <img src="https://res.cloudinary.com/cea/image/upload/h_120,q_auto,f_auto/ea-logo-square-1200x1200__1_.png" />}
+            </Link>
+            <div className={classes.localGroupContent} style={rowStyle}>
+              <div className={classes.localGroupNameRow}>
+                <Link to={`/groups/${group._id}`} className={classes.localGroupName}>{group.name}</Link>
+                <div className={classes.localGroupDistance}>
+                  {userLocation.known && group.mongoLocation?.coordinates && `${distance(group.mongoLocation.coordinates[1], group.mongoLocation.coordinates[0])} km`}
+                </div>
+              </div>
+              <div className={classes.localGroupLocation}>{group.location}</div>
+            </div>
+          </div>
+        })}
+      </div>
+      <div className={classes.localGroupsMap}>
+        <CommunityMapWrapper
+          mapOptions={userLocation.known ? {center: userLocation, zoom: 5} : {zoom: 1}}
+          hideLegend
+          showUsers={false}
+        />
+      </div>
+    </div>
+  )
+}
+
+const LocalGroupsComponent = registerComponent('LocalGroups', LocalGroups, {styles});
+
+declare global {
+  interface ComponentTypes {
+    LocalGroups: typeof LocalGroupsComponent
+  }
+}
