+++ packages/lesswrong/lib/components.ts
@@ -272,6 +272,8 @@ importComponent("EventCards", () => require('../components/events/modules/EventC
 importComponent("VirtualProgramCard", () => require('../components/events/modules/VirtualProgramCard'));
 // this is the new Community page, used by the EA Forum
 importComponent("Community", () => require('../components/community/Community'));
+importComponent("LocalGroups", () => require('../components/community/modules/LocalGroups'));
+importComponent("OnlineGroups", () => require('../components/community/modules/OnlineGroups'));
 // this is the previous Community page, used by LW
 importComponent("CommunityHome", () => require('../components/localGroups/CommunityHome'));
 importComponent(["CommunityMap", "PersonalMapLocationMarkers"], () => require('../components/localGroups/CommunityMap'));
